requirements.txt
weaviate-client
pandas
mysql-connector-python